+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 - [#4648](https://github.com/influxdb/influxdb/pull/4648): UDP Client (v2 client)
 - [#4690](https://github.com/influxdb/influxdb/pull/4690): SHOW SHARDS now includes database and policy. Thanks @pires
 - [#4676](https://github.com/influxdb/influxdb/pull/4676): UDP service listener performance enhancements
+- [#4659](https://github.com/influxdb/influxdb/pull/4659): Support IF EXISTS for DROP DATABASE
 
 ### Bugfixes
 - [#4643](https://github.com/influxdb/influxdb/pull/4643): Fix panic during backup restoration. Thanks @oiooj